BSSN Restructure (#100)
* Restructure BSSN codegen library

* Update examples

* Fix superB example

* Update to allow import from

* Update examples to remove GR aliases
