+++ nrpy/infrastructures/BHaH/general_relativity/BSSN/enforce_detgammabar_equals_detgammahat.py
@@ -1,0 +1,125 @@
+"""
+Generate C function to enforce the det(gammabar) = det(gammahat) constraint.
+
+Author: Zachariah B. Etienne
+        zachetie **at** gmail **dot* com
+"""
+
+from inspect import currentframe as cfr
+from types import FrameType as FT
+from typing import List, Union, cast
+
+import sympy as sp
+
+import nrpy.c_codegen as ccg
+import nrpy.c_function as cfc
+import nrpy.finite_difference as fin
+import nrpy.grid as gri
+import nrpy.helpers.parallel_codegen as pcg
+import nrpy.indexedexp as ixp
+import nrpy.infrastructures.BHaH.simple_loop as lp
+import nrpy.reference_metric as refmetric
+from nrpy.equations.general_relativity.BSSN_quantities import BSSN_quantities
+
+
+def register_CFunction_enforce_detgammabar_equals_detgammahat(
+    CoordSystem: str,
+    enable_rfm_precompute: bool,
+    enable_fd_functions: bool,
+    OMP_collapse: int,
+) -> Union[None, pcg.NRPyEnv_type]:
+    """
+    Register the function that enforces the det(gammabar) = det(gammahat) constraint.
+
+    :param CoordSystem: The coordinate system to be used.
+    :param enable_rfm_precompute: Whether to enable reference metric precomputation.
+    :param enable_fd_functions: Whether to enable finite difference functions.
+    :param OMP_collapse: Degree of OpenMP loop collapsing.
+
+    :return: None if in registration phase, else the updated NRPy environment.
+    """
+    if pcg.pcg_registration_phase():
+        pcg.register_func_call(f"{__name__}.{cast(FT, cfr()).f_code.co_name}", locals())
+        return None
+
+    Bq = BSSN_quantities[
+        CoordSystem + ("_rfm_precompute" if enable_rfm_precompute else "")
+    ]
+    rfm = refmetric.reference_metric[
+        CoordSystem + "_rfm_precompute" if enable_rfm_precompute else CoordSystem
+    ]
+
+    includes = ["BHaH_defines.h"]
+    desc = r"""Enforce det(gammabar) = det(gammahat) constraint. Required for strong hyperbolicity."""
+    cfunc_type = "void"
+    name = "enforce_detgammabar_equals_detgammahat"
+    params = "const commondata_struct *restrict commondata, const params_struct *restrict params, REAL *restrict xx[3], REAL *restrict in_gfs"
+    if enable_rfm_precompute:
+        params = params.replace(
+            "REAL *restrict xx[3]", "const rfm_struct *restrict rfmstruct"
+        )
+
+    # First define the Kronecker delta:
+    KroneckerDeltaDD = ixp.zerorank2()
+    for i in range(3):
+        KroneckerDeltaDD[i][i] = sp.sympify(1)
+
+    # The detgammabar in BSSN_RHSs is set to detgammahat when BSSN_RHSs::detgbarOverdetghat_equals_one=True (default),
+    #    so we manually compute it here:
+    dummygammabarUU, detgammabar = ixp.symm_matrix_inverter3x3(Bq.gammabarDD)
+
+    # Next apply the constraint enforcement equation above.
+    nrpyAbs = sp.Function("nrpyAbs")
+    hprimeDD = ixp.zerorank2()
+    for i in range(3):
+        for j in range(3):
+            hprimeDD[i][j] = (nrpyAbs(rfm.detgammahat) / detgammabar) ** (
+                sp.Rational(1, 3)
+            ) * (KroneckerDeltaDD[i][j] + Bq.hDD[i][j]) - KroneckerDeltaDD[i][j]
+
+    hDD_access_gfs: List[str] = []
+    hprimeDD_expr_list: List[sp.Expr] = []
+    for i in range(3):
+        for j in range(i, 3):
+            hDD_access_gfs += [
+                gri.BHaHGridFunction.access_gf(
+                    f"hDD{i}{j}", 0, 0, 0, gf_array_name="in_gfs"
+                )
+            ]
+            hprimeDD_expr_list += [hprimeDD[i][j]]
+
+    # To evaluate the cube root, SIMD support requires e.g., SLEEF.
+    #   Also need to be careful to not access memory out of bounds!
+    #   After all this is a loop over ALL POINTS.
+    #   Exercise to the reader: prove that for any reasonable grid,
+    #   SIMD loops over grid interiors never write data out of bounds
+    #   and are threadsafe for any reasonable number of threads.
+    body = lp.simple_loop(
+        loop_body=ccg.c_codegen(
+            hprimeDD_expr_list,
+            hDD_access_gfs,
+            enable_fd_codegen=True,
+            enable_simd=False,
+            enable_fd_functions=enable_fd_functions,
+        ),
+        loop_region="all points",
+        enable_intrinsics=False,
+        CoordSystem=CoordSystem,
+        enable_rfm_precompute=enable_rfm_precompute,
+        read_xxs=not enable_rfm_precompute,
+        OMP_collapse=OMP_collapse,
+    )
+
+    cfc.register_CFunction(
+        include_CodeParameters_h=True,
+        prefunc=fin.construct_FD_functions_prefunc() if enable_fd_functions else "",
+        includes=includes,
+        desc=desc,
+        cfunc_type=cfunc_type,
+        CoordSystem_for_wrapper_func=CoordSystem,
+        name=name,
+        params=params,
+        body=body,
+        enable_simd=False,
+    )
+    return cast(pcg.NRPyEnv_type, pcg.NRPyEnv())
